Add missing DeliveryStatus class
